refactor: Move registration config to its domain
