+++ internal/pkg/service/buffer/sink/tablesink/storage/volume/registration/registration.go
@@ -1,4 +1,4 @@
-package volume
+package registration
 
 import (
 	"context"
@@ -8,8 +8,7 @@ import (
 	"go.etcd.io/etcd/client/v3/concurrency"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/log"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/config"
-	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage"
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/sink/tablesink/storage/volume"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdop"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/etcdop/op"
 	"github.com/keboola/keboola-as-code/internal/pkg/service/common/servicectx"
@@ -19,18 +18,16 @@ type dependencies interface {
 	Logger() log.Logger
 	Process() *servicectx.Process
 	EtcdClient() *etcd.Client
-	Config() config.Config
 }
 
-type putOpFactory func(metadata storage.VolumeMetadata, id etcd.LeaseID) op.WithResult[storage.VolumeMetadata]
+type putOpFactory func(metadata volume.Metadata, id etcd.LeaseID) op.WithResult[volume.Metadata]
 
 // RegisterVolumes in etcd with lease, so on node failure, records are automatically removed after TTL seconds.
 // On session failure, volumes are registered again by the callback.
 // List of the active volumes can be read by the repository.VolumeRepository.
-func RegisterVolumes[V storage.Volume](d dependencies, volumes *Collection[V], putOpFactory putOpFactory) error {
+func RegisterVolumes[V volume.Volume](d dependencies, cfg Config, volumes *volume.Collection[V], putOpFactory putOpFactory) error {
 	logger := d.Logger()
 	client := d.EtcdClient()
-	cfg := d.Config()
 
 	// Graceful shutdown
 	ctx, cancel := context.WithCancel(context.Background())
@@ -43,8 +40,7 @@ func RegisterVolumes[V storage.Volume](d dependencies, volumes *Collection[V], p
 	})
 
 	// Register volumes
-	ttlSeconds := cfg.Sink.Table.Storage.Local.Volumes.RegistrationTTLSeconds
-	errCh := etcdop.ResistantSession(ctx, wg, logger, client, ttlSeconds, func(session *concurrency.Session) error {
+	errCh := etcdop.ResistantSession(ctx, wg, logger, client, cfg.TTLSeconds, func(session *concurrency.Session) error {
 		txn := op.Txn(client)
 		for _, vol := range volumes.All() {
 			txn.Then(putOpFactory(vol.Metadata(), session.Lease()))
